Modify arrow queries to stream results via chan
Instead of obtaining all arrow records for a given query before
returning to the caller, accept a channel as input and push
records to the channel one-by-one. This should enable streaming
data all the way from persistent storage (parquet files on disk)
to the flight client as arrow records, reducing the memory footprint
on the server side for any given query.